Summarize and Display Regression Results

This function processes the results from the regmonkey function, creating and displaying regression models for each significant variable combination identified. It utilizes the modelsummary package to format and present these results clearly, allowing for an easy evaluation of the significance and impact of various predictors.

Usage

summary_reg(significant_results, data, coef_map_override = NULL)

Arguments

significant_results

A list of significant variable combinations returned by regmonkey. Each element of the list should contain the dependent variable, the independent variable, and the controls under which the significant result was found.

data

The data frame containing all the variables referenced in the regression formulas.

coef_map_override

An optional named vector to override the default mapping of coefficient names in the output. This can be used to provide more readable or specific names for the variables in the regression output. If not provided, the function will automatically extract variable names from the models.

Value

This function does not return anything; it directly outputs the regression results formatted by the modelsummary package to the R console or to an R Markdown document.

Examples

# Assuming `regmonkey` has been run and stored results in `monkey_results`
x1 <- rnorm(100, mean = 0, sd = 1)
x2 <- rnorm(100, mean = 0, sd = 1)
x3 <- rnorm(100, mean = 0, sd = 1)
y <- 2 + 1.5 * x1 - 2 * x2 + 0.5 * x3 + rnorm(100, mean = 0, sd = 1)
data <- data.frame(y = y, x1 = x1, x2 = x2, x3 = x3)
significant_results <- regmonkey(data, "y", c("x1", "x2"), control_vars = c("x3"))
#> 
#> Processing independent variable: x1 
#> Processing combination 1 of 2 :  
#> Processing combination 2 of 2 : x3 
#> 
#> Processing independent variable: x2 
#> Processing combination 1 of 2 :  
#> Processing combination 2 of 2 : x3 
summary_reg(significant_results, data)
#> 
#> +-------------+----------+----------+-----------+-----------+
#> |             | (1)      | (2)      | (3)       | (4)       |
#> +=============+==========+==========+===========+===========+
#> | (Intercept) | 2.105*** | 2.174*** | 1.699***  | 1.835***  |
#> +-------------+----------+----------+-----------+-----------+
#> |             | (0.232)  | (0.230)  | (0.192)   | (0.176)   |
#> +-------------+----------+----------+-----------+-----------+
#> | x1          | 1.660*** | 1.596*** |           |           |
#> +-------------+----------+----------+-----------+-----------+
#> |             | (0.222)  | (0.221)  |           |           |
#> +-------------+----------+----------+-----------+-----------+
#> | x3          |          | 0.488*   |           | 0.846***  |
#> +-------------+----------+----------+-----------+-----------+
#> |             |          | (0.231)  |           | (0.177)   |
#> +-------------+----------+----------+-----------+-----------+
#> | x2          |          |          | -2.186*** | -2.237*** |
#> +-------------+----------+----------+-----------+-----------+
#> |             |          |          | (0.199)   | (0.180)   |
#> +-------------+----------+----------+-----------+-----------+
#> | Num.Obs.    | 100      | 100      | 100       | 100       |
#> +-------------+----------+----------+-----------+-----------+
#> | R2          | 0.362    | 0.390    | 0.552     | 0.637     |
#> +-------------+----------+----------+-----------+-----------+
#> | R2 Adj.     | 0.356    | 0.378    | 0.547     | 0.630     |
#> +-------------+----------+----------+-----------+-----------+
#> | AIC         | 452.8    | 450.3    | 417.6     | 398.4     |
#> +-------------+----------+----------+-----------+-----------+
#> | BIC         | 460.6    | 460.7    | 425.4     | 408.8     |
#> +-------------+----------+----------+-----------+-----------+
#> | Log.Lik.    | -223.400 | -221.147 | -205.800  | -195.183  |
#> +-------------+----------+----------+-----------+-----------+
#> | RMSE        | 2.26     | 2.21     | 1.89      | 1.70      |
#> +=============+==========+==========+===========+===========+
#> | + p < 0.1, * p < 0.05, ** p < 0.01, *** p < 0.001         |
#> +=============+==========+==========+===========+===========+